refactor: move ZeroDonationsSection to ProfileScreen
