+++ src/screens/users/ProfileScreen/CommunityDonationsImpactCards/ImpactDonationsVector.tsx
