--- src/screens/users/ProfileScreen/DirectDonationsImpactCards/ZeroDonationsSection/index.tsx
@@ -1,35 +1,0 @@
-import { theme } from "@ribon.io/shared";
-import Button from "components/atomics/buttons/Button";
-import { Text, View } from "react-native";
-import S from "./styles";
-
-type Props = {
-  title: string;
-  description: string;
-  buttonText: string;
-  onButtonPress: () => void;
-  image?: JSX.Element;
-};
-function ZeroDonationsSection({
-  title,
-  description,
-  buttonText,
-  onButtonPress,
-  image,
-}: Props): JSX.Element {
-  return (
-    <View style={S.zeroDonationsSection}>
-      {image}
-      <Text style={S.zeroDonationsTitle}>{title}</Text>
-      <Text style={S.zeroDonationsDescription}>{description}</Text>
-      <Button
-        text={buttonText}
-        onPress={onButtonPress}
-        textColor={theme.colors.green40}
-        customStyles={{ width: 200 }}
-      />
-    </View>
-  );
-}
-
-export default ZeroDonationsSection;
